Adds support for authentication via username and password
It's generally bad practice to expose endpoints to the internet without
some sort of authentication. Amazon SNS supports both Basic and Digest
authentication schemes. Basic is the easiest to implement but should
only be used with HTTPS since the credentials are easily converted to
plain text. Digest is more complicated to implement but can be
transmitted over HTTP without giving away the credentials.

Generally you should be using HTTPS if you care about the messages that
are being sent or received.